Rename `store` directory to `stores`
