+++ src/app/stores/backgroundStore.js
