+++ src/app/stores/panelStore.js
