+++ src/app/stores/windowStore.js
